+++ src/main/java/de/alpharogroup/file/system/SystemFileExtensions.java
@@ -70,6 +70,34 @@ public final class SystemFileExtensions
 	}
 
 	/**
+	 * Gets the users downloads directory from the system as {@link File} object
+	 *
+	 * @return the users downloads directory from the system as {@link File} object
+	 */
+	public static File getUserDownloadsDir()
+	{
+		return getUserDownloadsDir("");
+	}
+
+	/**
+	 * Gets the users downloads directory from the system as {@link File} object
+	 *
+	 * @param downloadsDirname
+	 * 			The name of the downloads directory, if null or empty the default value '/Downloads' will be taken
+	 *
+	 * @return the users downloads directory from the system as {@link File} object
+	 */
+	public static File getUserDownloadsDir(String downloadsDirname)
+	{
+		String ddn = downloadsDirname;
+		if(downloadsDirname == null || downloadsDirname.isEmpty()){
+			ddn = "/Downloads";
+		}
+		String userHomePath = System.getProperty("user.home");
+		return new File(userHomePath + ddn);
+	}
+
+	/**
 	 * Gets the user working directory from the system as {@link File} object
 	 *
 	 * @return the user working directory from the system as {@link File} object